Separte example app to have multiple example widgets
